+++ matchbox-server/src/main/java/ch/ahdis/matchbox/interceptors/MappingLanguageInterceptor.java
@@ -24,28 +24,41 @@ import static org.apache.commons.lang3.StringUtils.defaultString;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.util.Collections;
 import java.util.StringTokenizer;
 
+import ca.uhn.fhir.context.FhirVersionEnum;
+import ca.uhn.fhir.rest.server.exceptions.InternalErrorException;
+import ca.uhn.fhir.rest.server.exceptions.UnclassifiedServerFailureException;
+import ch.ahdis.matchbox.engine.MatchboxEngine;
+import ch.ahdis.matchbox.engine.exception.MatchboxUnsupportedFhirVersionException;
+import ch.ahdis.matchbox.mappinglanguage.MatchboxStructureMapUtilities;
+import ch.ahdis.matchbox.mappinglanguage.TransformSupportServices;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import org.hl7.fhir.convertors.factory.VersionConvertorFactory_40_50;
 import org.hl7.fhir.convertors.factory.VersionConvertorFactory_43_50;
-import org.hl7.fhir.exceptions.FHIRException;
 import org.hl7.fhir.instance.model.api.IIdType;
-import org.hl7.fhir.r5.model.StructureMap;
-import org.hl7.fhir.r5.model.Enumerations.FHIRVersion;
+import org.hl7.fhir.r5.context.IWorkerContext;
+import org.hl7.fhir.r5.model.Narrative;
 
 import ca.uhn.fhir.rest.api.Constants;
 import ca.uhn.fhir.rest.api.server.RequestDetails;
 import ca.uhn.fhir.rest.server.exceptions.AuthenticationException;
 import ca.uhn.fhir.rest.server.interceptor.InterceptorAdapter;
 import ch.ahdis.matchbox.util.MatchboxEngineSupport;
-import ch.ahdis.matchbox.engine.MatchboxEngine;
 import ch.ahdis.matchbox.spring.boot.autoconfigure.MutableHttpServletRequest;
+import org.hl7.fhir.r5.model.StructureMap;
+import org.hl7.fhir.r5.utils.structuremap.StructureMapUtilities;
+import org.hl7.fhir.utilities.xhtml.NodeType;
+import org.hl7.fhir.utilities.xhtml.XhtmlNode;
+
+import javax.annotation.Nullable;
 
 /**
- * MappingLanguagerInterceptor converts a FHIR Mapping Language texture
+ * MappingLanguageInterceptor converts a FHIR Mapping Language texture
  * representation in a StructureMap resource
  * 
  * POST {{host}}/$convert HTTP/1.1 Accept: application/fhir+json Content-Type:
@@ -58,81 +71,142 @@ import ch.ahdis.matchbox.spring.boot.autoconfigure.MutableHttpServletRequest;
  */
 public class MappingLanguageInterceptor extends InterceptorAdapter {
 
-	protected MatchboxEngineSupport matchboxEngineSupport;
+	protected final MatchboxEngineSupport matchboxEngineSupport;
 
 	private static final org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(MappingLanguageInterceptor.class);
 
-	public MappingLanguageInterceptor(MatchboxEngineSupport matchboxEngineSupport) {
+	public MappingLanguageInterceptor(final MatchboxEngineSupport matchboxEngineSupport) {
 		super();
 		this.matchboxEngineSupport = matchboxEngineSupport;
 	}
 
-	private FHIRVersion extractFhirVersion(String header) {
-		if (header == null) {
-			return null;
-		}
-		StringTokenizer tok = new StringTokenizer(header, ";");
-		String wantVersionString = null;
-		while (tok.hasMoreTokens()) {
-			String next = tok.nextToken().trim();
-			if (next.startsWith("fhirVersion=")) {
-				wantVersionString = next.substring("fhirVersion=".length()).trim();
-				break;
-			}
-		}
-		if (isNotBlank(wantVersionString)) {
-			return FHIRVersion.fromCode(wantVersionString);
-		}
-		return FHIRVersion.fromCode(matchboxEngineSupport.getClientContext().getFhirVersion());
-	}
-
-	public StructureMap parseMap(String content) throws FHIRException {
-		MatchboxEngine matchboxEngine = matchboxEngineSupport.getMatchboxEngine("default", null, true, false);
-		return matchboxEngine.parseMapR5(content);
-	}
-
 	@Override
-	public boolean incomingRequestPostProcessed(RequestDetails theRequestDetails, HttpServletRequest theRequest,
-			HttpServletResponse theResponse) throws AuthenticationException {
+	public boolean incomingRequestPostProcessed(final RequestDetails theRequestDetails,
+															  final HttpServletRequest theRequest,
+															  final HttpServletResponse theResponse) throws AuthenticationException {
 
-		String contentType = defaultString(theRequest.getHeader(Constants.HEADER_CONTENT_TYPE));
+		final String contentType = defaultString(theRequest.getHeader(Constants.HEADER_CONTENT_TYPE));
 		if (contentType.startsWith("text/fhir-mapping")) {
 			log.debug("processing text/fhir mapping - converting to json");
-			FHIRVersion fhirVersion = this.extractFhirVersion(contentType);
-			((MutableHttpServletRequest) theRequest).putHeader(Constants.HEADER_CONTENT_TYPE, "application/fhir+json;fhirVersion="+fhirVersion.getDisplay());
 
-			
-			StructureMap structureMap = parseMap(new String(theRequestDetails.loadRequestContents()));
+			FhirVersionEnum requestedFhirVersion = this.extractFhirVersion(contentType);
+			final FhirVersionEnum mainEngineFhirVersion =
+				FhirVersionEnum.forVersionString(this.matchboxEngineSupport.getClientContext().getFhirVersion());
+
+			if (requestedFhirVersion != null && this.matchboxEngineSupport.getClientContext().getOnlyOneEngine()) {
+				// If the onlyOneEngine mode is enabled, check that it's the same FHIR version
+				if (mainEngineFhirVersion != requestedFhirVersion) {
+					throw new UnclassifiedServerFailureException(415, "The requested FHIR version (" + requestedFhirVersion +
+						") does not match the engine's FHIR version (" + mainEngineFhirVersion + ")");
+				}
+			} else if (requestedFhirVersion == null) {
+				// If the requested FHIR version is not specified, use the engine's FHIR version
+				requestedFhirVersion = mainEngineFhirVersion;
+			}
+
+			// Update the MIME type in the request
+			((MutableHttpServletRequest) theRequest).putHeader(Constants.HEADER_CONTENT_TYPE, "application/fhir+json;" +
+				"fhirVersion="+requestedFhirVersion.getFhirVersionString());
+
+			final var mapText = new String(theRequestDetails.loadRequestContents(), StandardCharsets.UTF_8);
+
+			// Force the StructureMap ID to a specific value for PUT requests
+			String structureMapId = null;
 			if ("PUT".equals(theRequest.getMethod())) {
-				IIdType id = theRequestDetails.getId();
+				final IIdType id = theRequestDetails.getId();
 				if (id != null) {
-					structureMap.setId(id.getIdPart());
+					structureMapId = id.getIdPart();
 				}
 			}
 
-			String json = null;
+			final String jsonStructureMap;
 			try {
-				switch(fhirVersion) {
-					case _4_0_1:
-						org.hl7.fhir.r4.model.StructureMap sm4 =  (org.hl7.fhir.r4.model.StructureMap) VersionConvertorFactory_40_50.convertResource(structureMap);
-						json = new org.hl7.fhir.r4.formats.JsonParser().composeString(sm4);				
-						break;
-					case _4_3_0:
-						org.hl7.fhir.r4b.model.StructureMap sm4b =  (org.hl7.fhir.r4b.model.StructureMap) VersionConvertorFactory_43_50.convertResource(structureMap);
-						json = new org.hl7.fhir.r4b.formats.JsonParser().composeString(sm4b);				
-						break;
-					case _5_0_0:
-						json = new org.hl7.fhir.r5.formats.JsonParser().composeString(structureMap);				
-						break;
-					default:
-						log.error("FHIR Version not supported "+fhirVersion.getDisplay());
-				}
-				theRequestDetails.setRequestContents(json.getBytes());
-			} catch (IOException e) {
-				log.error("Exception while conversion to json", e);
+				jsonStructureMap = this.compileAndSerializeMap(mapText, requestedFhirVersion, structureMapId);
+			} catch (final IOException e) {
+				// Nothing we can do here, propagate the exception (error 500)
+				throw new InternalErrorException("Failed to convert StructureMap to JSON", e);
 			}
+			theRequestDetails.setRequestContents(jsonStructureMap.getBytes(StandardCharsets.UTF_8));
 		}
 		return true;
 	}
 
+	@Nullable
+	private FhirVersionEnum extractFhirVersion(final String header) {
+		if (header != null) {
+			final var tokenizer = new StringTokenizer(header, ";");
+			String wantVersionString = null;
+			while (tokenizer.hasMoreTokens()) {
+				String next = tokenizer.nextToken().trim();
+				if (next.startsWith("fhirVersion=")) {
+					wantVersionString = next.substring("fhirVersion=".length()).trim();
+					break;
+				}
+			}
+			if (isNotBlank(wantVersionString)) {
+				return FhirVersionEnum.forVersionString(wantVersionString);
+			}
+		}
+		// Returns null if no FHIR version can be extracted from the header
+		return null;
+	}
+
+	/**
+	 * Compiles the given FHIR Mapping Language text and serializes the resulting StructureMap resource. The
+	 * compilation always happens in the FHIR R5 version, then the resulting resource is converted to the requested FHIR
+	 * version.
+	 *
+	 * @param mapText The FHIR map as text.
+	 * @param fhirVersion The FHIR version to use for the serialization.
+	 * @param structureMapId The id to set on the StructureMap resource, or null if no id should be set.
+	 * @return
+	 * @throws IOException
+	 */
+	public String compileAndSerializeMap(final String mapText,
+													 final FhirVersionEnum fhirVersion,
+													 final @Nullable String structureMapId) throws IOException {
+		final IWorkerContext workerContext;
+		if (this.matchboxEngineSupport.getClientContext().getOnlyOneEngine()) {
+			// In 'onlyOneEngine' mode, we cannot initialize a new engine for the compilation
+			if (FhirVersionEnum.R5.getFhirVersionString().equals(this.matchboxEngineSupport.getClientContext().getFhirVersion())) {
+				// We try to reuse the main engine if it's R5
+				workerContext = MatchboxEngineSupport.mainEngine.getContext();
+			} else {
+				// Otherwise, we have to initialize a new worker context
+				workerContext = MatchboxEngine.createR5WorkerContext();
+			}
+		} else {
+			// Let's get or initialize a new R5 engine
+			workerContext = MatchboxEngineSupport.mainEngine.getContextForFhirVersion(fhirVersion.getFhirVersionString());
+		}
+
+		final var scu = new MatchboxStructureMapUtilities(workerContext,
+																		  new TransformSupportServices(workerContext,
+																												 Collections.emptyList()),
+																		  MatchboxEngineSupport.mainEngine);
+		final StructureMap mapR5 = scu.parse(mapText, "map");
+		mapR5.getText().setStatus(Narrative.NarrativeStatus.GENERATED);
+		mapR5.getText().setDiv(new XhtmlNode(NodeType.Element, "div"));
+		final String render = StructureMapUtilities.render(mapR5);
+		mapR5.getText().getDiv().addTag("pre").addText(render);
+		if (structureMapId != null) {
+			mapR5.setId(structureMapId);
+		}
+
+		return switch (fhirVersion) {
+			case R4:
+				org.hl7.fhir.r4.model.StructureMap sm4 = (org.hl7.fhir.r4.model.StructureMap) VersionConvertorFactory_40_50.convertResource(
+					mapR5);
+				yield new org.hl7.fhir.r4.formats.JsonParser().composeString(sm4);
+			case R4B:
+				org.hl7.fhir.r4b.model.StructureMap sm4b = (org.hl7.fhir.r4b.model.StructureMap) VersionConvertorFactory_43_50.convertResource(
+					mapR5);
+				yield new org.hl7.fhir.r4b.formats.JsonParser().composeString(sm4b);
+			case R5:
+				yield new org.hl7.fhir.r5.formats.JsonParser().composeString(mapR5);
+			default:
+				throw new MatchboxUnsupportedFhirVersionException("Unsupported FHIR version: " + fhirVersion,
+																				  fhirVersion);
+		};
+	}
 }