+++ docs/changelog.md
@@ -3,7 +3,8 @@ Unreleased
 - Upgrade hapifhir org.hl7.fhir.core to 6.4.4
 - Remove the `devMode` configuration parameter, it is now enabled when `httpReadOnly` is not [#315](https://github.com/ahdis/matchbox/issues/315)
 - Remove the `autoInstallMissingIgs` configuration parameter, it is now enabled when `httpReadOnly` is not [#315](https://github.com/ahdis/matchbox/issues/315)
-- Improved the Matchbox server documentation [#315](https://github.com/ahdis/matchbox/issues/315)
+- Improve the Matchbox server documentation [#315](https://github.com/ahdis/matchbox/issues/315)
+- Respect the 'onlyOneEngine' mode in MappingLanguageInterceptor [#316](https://github.com/ahdis/matchbox/issues/316)
 
 2024/11/25 Release 3.9.8